+++ test/runtests.jl
@@ -21,7 +21,13 @@ begin
     @safetestset "Symmetric Matrix             " include("special_matrices/symmetric_matrix.jl")
 end
 begin
+    @safetestset "Triangular Matrices          " include("special_matrices/triangular.jl")
+end
+begin
     @safetestset "Scalar mul! Return Value     " include("special_matrices/scalar_mul_return_value.jl")
+end
+begin
+    @safetestset "Optimizer Primitives         " include("special_matrices/optimizer_primitives.jl")
 end
 begin
     @safetestset "Grassmann Lie Alg Hor        " include("lie_algebras/grassmann_lie_algebra_horizontal.jl")